[All] Renamed io.vertigo.core.di to io.vertigo.core.component.di
