+++ vertigo-struts2/src/main/java/io/vertigo/struts2/core/AbstractActionSupport.java
@@ -20,7 +20,7 @@ package io.vertigo.struts2.core;
 
 import io.vertigo.commons.config.ConfigManager;
 import io.vertigo.core.Home;
-import io.vertigo.core.di.injector.Injector;
+import io.vertigo.core.component.di.injector.Injector;
 import io.vertigo.lang.Assertion;
 import io.vertigo.struts2.context.ContextCacheManager;
 import io.vertigo.struts2.exception.ExpiredContextException;
