+++ vertigo-struts2/src/main/java/io/vertigo/struts2/impl/servlet/HomeServletStarter.java
@@ -19,8 +19,7 @@
 package io.vertigo.struts2.impl.servlet;
 
 import io.vertigo.boot.xml.XMLAppConfigBuilder;
-import io.vertigo.core.Home.App;
-import io.vertigo.core.boot.BootConfigBuilder;
+import io.vertigo.core.App;
 import io.vertigo.core.config.LogConfig;
 import io.vertigo.lang.Assertion;
 import io.vertigo.struts2.plugins.config.servlet.WebAppContextConfigPlugin;
@@ -68,10 +67,10 @@ final class HomeServletStarter {
 			final Properties bootConf = createBootProperties(servletContext);
 			Assertion.checkArgument(bootConf.containsKey("boot.applicationConfiguration"), "Param \"boot.applicationConfiguration\" is mandatory, check your .properties or web.xml.");
 
-			final XMLAppConfigBuilder appConfigBuilder = new XMLAppConfigBuilder()
-					.withBootConfig(new BootConfigBuilder()
-							.silently()
-							.build());
+			final XMLAppConfigBuilder appConfigBuilder = new XMLAppConfigBuilder();
+			appConfigBuilder.beginBoot()
+					.silently()
+					.endBoot();
 
 			//si présent on récupère le paramétrage du fichier externe de paramétrage log4j
 			if (bootConf.containsKey(LOG4J_CONFIGURATION_PARAM_NAME)) {
