+++ vertigo-struts2/src/main/java/io/vertigo/struts2/plugins/resource/servlet/ServletResourceResolverPlugin.java
@@ -18,7 +18,7 @@
  */
 package io.vertigo.struts2.plugins.resource.servlet;
 
-import io.vertigo.commons.impl.resource.ResourceResolverPlugin;
+import io.vertigo.core.impl.resource.ResourceResolverPlugin;
 import io.vertigo.lang.Assertion;
 import io.vertigo.lang.Option;
